+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # pygdbmi release history
 
+## 0.10.0.0
+
+ **Breaking Changes**
+
+* Drop support for Python 3.5
+* Update `GdbController()` API. New API is `GdbController(command: Optional[List[str]], time_to_check_for_additional_output_sec: Optional[int])`.
+* `GdbController.verify_valid_gdb_subprocess()` was removed
+* Remove `NoGdbProcessError` error
+
+Other Changes
+
+* Add new `IoManager` class to handle more generic use-cases
+* [dev] use pytest for testing
+* gdb mi parsing remains unchanged
+
 ## 0.9.0.3
 
 * Drop support for 2.7, 3.4